+++ src/scalecast/util.py
@@ -1,8 +1,196 @@
 from scalecast.Forecaster import descriptive_assert, Forecaster
+from scalecast import Forecaster
 import pandas_datareader as pdr
 import matplotlib.pyplot as plt
 import seaborn as sns
 import numpy as np
+import pandas as pd
+
+class metrics:
+    def mape(a,f):
+        """ mean absolute percentage error (MAPE).
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived MAPE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.mape(a,f)
+        """
+        return Forecaster.mape(a,f)
+
+    def r2(a,f):
+        """ r-squared (R2).
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived R2.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.r2(a,f)
+        """
+        return Forecaster.r2(a,f)
+
+    def mse(a,f):
+        """ mean squared error (MSE).
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived MSE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.mse(a,f)
+        """
+        return Forecaster.rmse(a,f)**2
+
+    def rmse(a,f):
+        """ root mean squared error (RMSE).
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived RMSE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.rmse(a,f)
+        """
+        return Forecaster.rmse(a,f)
+
+    def mae(a,f):
+        """ mean absolute error (MAE).
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived MAE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.mae(a,f)
+        """
+        return Forecaster.mae(a,f)
+
+    def smape(a,f):
+        """ symmetric mean absolute percentage error (sMAPE).
+        uses the same definition as used in M4 competition.
+        does not multiply by 100.
+        see https://ideas.repec.org/a/eee/intfor/v36y2020i1p54-74.html.
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+
+        Returns:
+            (float): the derived sMAPE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> metrics.smape(a,f)
+        """
+        a = np.array(a)
+        f = np.array(f)
+        return (
+            1/len(a) *
+            np.sum(
+                2*np.abs(f-a) / (
+                    np.abs(a) + np.abs(f)
+                )
+            )
+        )
+
+    def mase(a,f,obs,m):
+        """ mean absolute scaled error (MASE).
+        uses the same definition as used in M4 competition.
+        see https://ideas.repec.org/a/eee/intfor/v36y2020i1p54-74.html.
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            f (list-like): the predictions over the forecast horizon.
+            obs (list-like): the actual observations used to create the forecast.
+            m (int): the seasonal period.
+
+        Returns:
+            (float): the derived MAPE.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> obs = [-5,-4,-3,-2,-1,0]
+        >>> metrics.mase(a,f,obs,1)
+        """
+        a = np.array(a)
+        f = np.array(f)
+        avger = 1/len(a)
+        num = np.sum(np.abs(f-a))
+        davger = 1 / (len(obs) - m)
+        denom = np.sum(
+            np.abs(
+                pd.Series(obs).diff(m).values[m:]
+            )
+        )
+        return avger * (num / (davger * denom))
+
+    def msis(a,uf,lf,obs,m,alpha=0.05):
+        """ mean scaled interval score (MSIS) for evaluating confidence intervals.
+        uses the same definition as used in M4 competition.
+        lower values are better.
+        see https://ideas.repec.org/a/eee/intfor/v36y2020i1p54-74.html.
+
+        Args:
+            a (list-like): the actuals over the forecast horizon.
+            uf (list-like): the upper-forecast bound according to the confidence interval.
+            lf (list-like): the lower-forecast bound according to the confidence interval.
+            obs (list-like): the actual observations used to create the forecast.
+            m (int): the seasonal period.
+            alpha (float): default 0.05. 0.05 for 95% confidence intervals, etc.
+
+        Returns:
+            (float): the derived MSIS.
+
+        >>> from scalecast.util import metrics
+        >>> a = [1,2,3,4,5]
+        >>> f = [1,2,3,4,6]
+        >>> uf = [1.5,2.5,3.5,4.5,6.5]
+        >>> lf = [.5,1.5,2.5,3.5,5.5]
+        >>> obs = [-5,-4,-3,-2,-1,0]
+        >>> metrics.msis(a,uf,lf,obs,1) # returns a value of 5.0
+        """
+        a = np.array(a)
+        uf = np.array(uf)
+        lf = np.array(lf)
+        avger = 1/len(a)
+        num1 = uf-lf
+        num2 = np.array([(
+            (2/alpha*(lfs-acs)) if lfs > acs else
+            (2/alpha*(acs-ufs)) if acs > ufs else
+            0
+        ) for acs, ufs, lfs in zip(a,uf,lf)])
+        davger = 1/(len(obs) - m)
+        denom = np.sum(np.abs(pd.Series(obs).diff(m).values[m:]))
+        return avger * (np.sum(num1 + num2) / (davger * denom))
 
 def pdr_load(
     sym,